types: Fix onSuccess callback
Currently we have a `code` property on the `GoogleLoginResponse` interface that isn't actually there and is just a hack for people doing the offline flow. Additionally, if using the online flow in TypeScript you can't safely access any of the properties and need to cast the argument of the onSuccess callback to get them.

This PR fixes these issues by properly handling the `responseType` property that determines what the `onSuccess` callback's argument is.