Added ability to specify active, hover, disabled and icon inline styles
The implementation of the new style merging is to use a function, named 'addKeysTo' which takes 2 objects, copies a reference to the target object, and assigns each style from the user defined style object too it. The reason that I am using a function for this operation is to follow the DRY principle, and to minimize the bundle size.